move counter updating to background pool
