+++ execution/executor/src/metrics.rs
@@ -2,12 +2,23 @@
 // Parts of the project are originally copyright © Meta Platforms, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
+use aptos_logger::{prelude::*, sample, warn};
 use aptos_metrics_core::{
     exponential_buckets, register_histogram, register_histogram_vec, register_int_counter,
     register_int_counter_vec, register_int_gauge_vec, Histogram, HistogramVec, IntCounter,
     IntCounterVec, IntGaugeVec,
 };
+use aptos_types::{
+    contract_event::ContractEvent,
+    transaction::{
+        authenticator::AccountAuthenticator, signature_verified_transaction::TransactionProvider,
+        ExecutionStatus, Transaction, TransactionOutput, TransactionStatus,
+    },
+};
+use aptos_vm::AptosVM;
+use move_core_types::{language_storage::CORE_CODE_ADDRESS, vm_status::StatusCode};
 use once_cell::sync::Lazy;
+use std::time::Duration;
 
 pub static EXECUTE_CHUNK: Lazy<Histogram> = Lazy::new(|| {
     register_histogram!(
@@ -248,3 +259,245 @@ pub static CONCURRENCY_GAUGE: Lazy<IntGaugeVec> = Lazy::new(|| {
     )
     .unwrap()
 });
+
+pub fn update_counters_for_processed_chunk<T>(
+    transactions: &[T],
+    transaction_outputs: &[TransactionOutput],
+    process_type: &str,
+) where
+    T: TransactionProvider,
+{
+    let detailed_counters = AptosVM::get_processed_transactions_detailed_counters();
+    let detailed_counters_label = if detailed_counters { "true" } else { "false" };
+    if transactions.len() != transaction_outputs.len() {
+        warn!(
+            "Chunk lenthgs don't match: txns: {} and outputs: {}",
+            transactions.len(),
+            transaction_outputs.len()
+        );
+    }
+
+    for (txn, output) in transactions.iter().zip(transaction_outputs.iter()) {
+        if detailed_counters {
+            if let Ok(size) = bcs::serialized_size(output) {
+                PROCESSED_TXNS_OUTPUT_SIZE
+                    .with_label_values(&[process_type])
+                    .observe(size as f64);
+            }
+        }
+
+        let (state, reason, error_code) = match output.status() {
+            TransactionStatus::Keep(execution_status) => match execution_status {
+                ExecutionStatus::Success => ("keep_success", "", "".to_string()),
+                ExecutionStatus::OutOfGas => ("keep_rejected", "OutOfGas", "error".to_string()),
+                ExecutionStatus::MoveAbort { info, .. } => (
+                    "keep_rejected",
+                    "MoveAbort",
+                    if detailed_counters {
+                        info.as_ref()
+                            .map(|v| v.reason_name.to_lowercase())
+                            .unwrap_or_else(|| "none".to_string())
+                    } else {
+                        "error".to_string()
+                    },
+                ),
+                ExecutionStatus::ExecutionFailure { .. } => {
+                    ("keep_rejected", "ExecutionFailure", "error".to_string())
+                },
+                ExecutionStatus::MiscellaneousError(e) => (
+                    "keep_rejected",
+                    "MiscellaneousError",
+                    if detailed_counters {
+                        e.map(|v| format!("{:?}", v).to_lowercase())
+                            .unwrap_or_else(|| "none".to_string())
+                    } else {
+                        "error".to_string()
+                    },
+                ),
+            },
+            TransactionStatus::Discard(discard_status_code) => {
+                (
+                    // Specialize duplicate txns for alerts
+                    if *discard_status_code == StatusCode::SEQUENCE_NUMBER_TOO_OLD {
+                        "discard_sequence_number_too_old"
+                    } else if *discard_status_code == StatusCode::SEQUENCE_NUMBER_TOO_NEW {
+                        "discard_sequence_number_too_new"
+                    } else if *discard_status_code == StatusCode::TRANSACTION_EXPIRED {
+                        "discard_transaction_expired"
+                    } else {
+                        // Only log if it is an interesting discard
+                        sample!(
+                            SampleRate::Duration(Duration::from_secs(15)),
+                            warn!(
+                                "[sampled] Txn being discarded is {:?} with status code {:?}",
+                                txn, discard_status_code
+                            )
+                        );
+                        "discard"
+                    },
+                    "error_code",
+                    if detailed_counters {
+                        format!("{:?}", discard_status_code).to_lowercase()
+                    } else {
+                        "error".to_string()
+                    },
+                )
+            },
+            TransactionStatus::Retry => ("retry", "", "".to_string()),
+        };
+
+        let kind = match txn.get_transaction() {
+            Some(Transaction::UserTransaction(_)) => "user_transaction",
+            Some(Transaction::GenesisTransaction(_)) => "genesis",
+            Some(Transaction::BlockMetadata(_)) => "block_metadata",
+            Some(Transaction::BlockMetadataExt(_)) => "block_metadata_ext",
+            Some(Transaction::StateCheckpoint(_)) => "state_checkpoint",
+            Some(Transaction::BlockEpilogue(_)) => "block_epilogue",
+            Some(Transaction::ValidatorTransaction(_)) => "validator_transaction",
+            None => "unknown",
+        };
+
+        PROCESSED_TXNS_COUNT
+            .with_label_values(&[process_type, kind, state])
+            .inc();
+
+        if !error_code.is_empty() {
+            PROCESSED_FAILED_TXNS_REASON_COUNT
+                .with_label_values(&[
+                    detailed_counters_label,
+                    process_type,
+                    state,
+                    reason,
+                    &error_code,
+                ])
+                .inc();
+        }
+
+        if let Some(Transaction::UserTransaction(user_txn)) = txn.get_transaction() {
+            if detailed_counters {
+                let mut signature_count = 0;
+                let account_authenticators = user_txn.authenticator_ref().all_signers();
+                for account_authenticator in account_authenticators {
+                    match account_authenticator {
+                        AccountAuthenticator::Ed25519 { .. } => {
+                            signature_count += 1;
+                            PROCESSED_TXNS_AUTHENTICATOR
+                                .with_label_values(&[process_type, "Ed25519"])
+                                .inc();
+                        },
+                        AccountAuthenticator::MultiEd25519 { signature, .. } => {
+                            let count = signature.signatures().len();
+                            signature_count += count;
+                            PROCESSED_TXNS_AUTHENTICATOR
+                                .with_label_values(&[process_type, "Ed25519_in_MultiEd25519"])
+                                .inc_by(count as u64);
+                        },
+                        AccountAuthenticator::SingleKey { authenticator } => {
+                            signature_count += 1;
+                            PROCESSED_TXNS_AUTHENTICATOR
+                                .with_label_values(&[
+                                    process_type,
+                                    &format!("{}_in_SingleKey", authenticator.signature().name()),
+                                ])
+                                .inc();
+                        },
+                        AccountAuthenticator::MultiKey { authenticator } => {
+                            for (_, signature) in authenticator.signatures() {
+                                signature_count += 1;
+                                PROCESSED_TXNS_AUTHENTICATOR
+                                    .with_label_values(&[
+                                        process_type,
+                                        &format!("{}_in_MultiKey", signature.name()),
+                                    ])
+                                    .inc();
+                            }
+                        },
+                        AccountAuthenticator::NoAccountAuthenticator => {
+                            PROCESSED_TXNS_AUTHENTICATOR
+                                .with_label_values(&[process_type, "NoAccountAuthenticator"])
+                                .inc();
+                        },
+                    };
+                }
+
+                PROCESSED_TXNS_NUM_AUTHENTICATORS
+                    .with_label_values(&[process_type])
+                    .observe(signature_count as f64);
+            }
+
+            match user_txn.payload() {
+                aptos_types::transaction::TransactionPayload::Script(_script) => {
+                    PROCESSED_USER_TXNS_BY_PAYLOAD
+                        .with_label_values(&[process_type, "script", state])
+                        .inc();
+                },
+                aptos_types::transaction::TransactionPayload::EntryFunction(function) => {
+                    PROCESSED_USER_TXNS_BY_PAYLOAD
+                        .with_label_values(&[process_type, "function", state])
+                        .inc();
+
+                    let is_core = function.module().address() == &CORE_CODE_ADDRESS;
+                    PROCESSED_USER_TXNS_ENTRY_FUNCTION_BY_MODULE
+                        .with_label_values(&[
+                            detailed_counters_label,
+                            process_type,
+                            if is_core { "core" } else { "user" },
+                            if detailed_counters {
+                                function.module().name().as_str()
+                            } else if is_core {
+                                "core_module"
+                            } else {
+                                "user_module"
+                            },
+                            state,
+                        ])
+                        .inc();
+                    if is_core && detailed_counters {
+                        PROCESSED_USER_TXNS_ENTRY_FUNCTION_BY_CORE_METHOD
+                            .with_label_values(&[
+                                process_type,
+                                function.module().name().as_str(),
+                                function.function().as_str(),
+                                state,
+                            ])
+                            .inc();
+                    }
+                },
+                aptos_types::transaction::TransactionPayload::Multisig(_) => {
+                    PROCESSED_USER_TXNS_BY_PAYLOAD
+                        .with_label_values(&[process_type, "multisig", state])
+                        .inc();
+                },
+
+                // Deprecated.
+                aptos_types::transaction::TransactionPayload::ModuleBundle(_) => {
+                    PROCESSED_USER_TXNS_BY_PAYLOAD
+                        .with_label_values(&[process_type, "deprecated_module_bundle", state])
+                        .inc();
+                },
+            }
+        }
+
+        for event in output.events() {
+            let (is_core, creation_number) = match event {
+                ContractEvent::V1(v1) => (
+                    v1.key().get_creator_address() == CORE_CODE_ADDRESS,
+                    if detailed_counters {
+                        v1.key().get_creation_number().to_string()
+                    } else {
+                        "event".to_string()
+                    },
+                ),
+                ContractEvent::V2(_v2) => (false, "event".to_string()),
+            };
+            PROCESSED_USER_TXNS_CORE_EVENTS
+                .with_label_values(&[
+                    detailed_counters_label,
+                    process_type,
+                    if is_core { "core" } else { "user" },
+                    &creation_number,
+                ])
+                .inc();
+        }
+    }
+}
