+++ execution/executor/src/workflow/do_get_execution_output.rs
@@ -15,6 +15,7 @@ use aptos_executor_types::{
     execution_output::ExecutionOutput, should_forward_to_subscription_service,
     transactions_with_output::TransactionsWithOutput,
 };
+use aptos_experimental_runtimes::thread_manager::THREAD_MANAGER;
 use aptos_logger::prelude::*;
 use aptos_metrics_core::TimerHelper;
 use aptos_storage_interface::cached_state_view::{CachedStateView, StateCache};
@@ -23,7 +24,6 @@ use aptos_types::{
         config::BlockExecutorConfigFromOnchain,
         partitioner::{ExecutableTransactions, PartitionedTransactions},
     },
-    contract_event::ContractEvent,
     epoch_state::EpochState,
     on_chain_config::{ConfigurationResource, OnChainConfig, ValidatorSet},
     state_store::{
@@ -31,17 +31,14 @@ use aptos_types::{
         TStateView,
     },
     transaction::{
-        authenticator::AccountAuthenticator,
-        signature_verified_transaction::{SignatureVerifiedTransaction, TransactionProvider},
-        BlockEndInfo, BlockOutput, ExecutionStatus, Transaction, TransactionOutput,
-        TransactionStatus, Version,
+        signature_verified_transaction::SignatureVerifiedTransaction, BlockEndInfo, BlockOutput,
+        Transaction, TransactionOutput, TransactionStatus, Version,
     },
     write_set::{TransactionWrite, WriteSet},
 };
-use aptos_vm::{AptosVM, VMExecutor};
+use aptos_vm::VMExecutor;
 use itertools::Itertools;
-use move_core_types::{language_storage::CORE_CODE_ADDRESS, vm_status::StatusCode};
-use std::{iter, sync::Arc, time::Duration};
+use std::{iter, sync::Arc};
 
 pub struct DoGetExecutionOutput;
 
@@ -52,7 +49,7 @@ impl DoGetExecutionOutput {
         onchain_config: BlockExecutorConfigFromOnchain,
         append_state_checkpoint_to_block: Option<HashValue>,
     ) -> Result<ExecutionOutput> {
-        let res = match transactions {
+        let out = match transactions {
             ExecutableTransactions::Unsharded(txns) => {
                 Self::by_transaction_execution_unsharded::<V>(
                     txns,
@@ -69,14 +66,19 @@ impl DoGetExecutionOutput {
             )?,
         };
 
-        {
-            let _timer = OTHER_TIMERS.timer_with(&["update_counters__by_execution"]);
-            for x in [&res.to_commit, &res.to_discard, &res.to_retry] {
-                update_counters_for_processed_chunk(x.txns(), x.transaction_outputs(), "execution");
+        let ret = out.clone();
+        THREAD_MANAGER.get_background_pool().spawn(move || {
+            let _timer = OTHER_TIMERS.timer_with(&["async_update_counters__by_execution"]);
+            for x in [&out.to_commit, &out.to_retry, &out.to_discard] {
+                metrics::update_counters_for_processed_chunk(
+                    &x.transactions,
+                    &x.transaction_outputs,
+                    "execution",
+                )
             }
-        }
+        });
 
-        Ok(res)
+        Ok(ret)
     }
 
     fn by_transaction_execution_unsharded<V: VMExecutor>(
@@ -134,8 +136,6 @@ impl DoGetExecutionOutput {
         transaction_outputs: Vec<TransactionOutput>,
         state_view: CachedStateView,
     ) -> Result<ExecutionOutput> {
-        update_counters_for_processed_chunk(&transactions, &transaction_outputs, "output");
-
         // collect all accounts touched and dedup
         let write_set = transaction_outputs
             .iter()
@@ -145,14 +145,26 @@ impl DoGetExecutionOutput {
         // prime the state cache by fetching all touched accounts
         state_view.prime_cache_by_write_set(write_set)?;
 
-        Parser::parse(
+        let out = Parser::parse(
             state_view.next_version(),
             transactions,
             transaction_outputs,
             state_view.into_state_cache(),
             None, // block end info
             None, // append state checkpoint to block
-        )
+        )?;
+
+        let ret = out.clone();
+        THREAD_MANAGER.get_background_pool().spawn(move || {
+            let _timer = OTHER_TIMERS.timer_with(&["async_update_counters__by_output"]);
+            metrics::update_counters_for_processed_chunk(
+                &out.to_commit.transactions,
+                &out.to_commit.transaction_outputs,
+                "output",
+            )
+        });
+
+        Ok(ret)
     }
 
     fn execute_block_sharded<V: VMExecutor>(
@@ -185,6 +197,7 @@ impl DoGetExecutionOutput {
         state_view: &CachedStateView,
         onchain_config: BlockExecutorConfigFromOnchain,
     ) -> Result<BlockOutput<TransactionOutput>> {
+        let _timer = OTHER_TIMERS.timer_with(&["vm_execute_block"]);
         Ok(V::execute_block(transactions, state_view, onchain_config)?)
     }
 
@@ -225,248 +238,6 @@ impl DoGetExecutionOutput {
             ),
         };
         Ok(transaction_outputs)
-    }
-}
-
-pub fn update_counters_for_processed_chunk<T>(
-    transactions: &[T],
-    transaction_outputs: &[TransactionOutput],
-    process_type: &str,
-) where
-    T: TransactionProvider,
-{
-    let detailed_counters = AptosVM::get_processed_transactions_detailed_counters();
-    let detailed_counters_label = if detailed_counters { "true" } else { "false" };
-    if transactions.len() != transaction_outputs.len() {
-        warn!(
-            "Chunk lenthgs don't match: txns: {} and outputs: {}",
-            transactions.len(),
-            transaction_outputs.len()
-        );
-    }
-
-    for (txn, output) in transactions.iter().zip(transaction_outputs.iter()) {
-        if detailed_counters {
-            if let Ok(size) = bcs::serialized_size(output) {
-                metrics::PROCESSED_TXNS_OUTPUT_SIZE
-                    .with_label_values(&[process_type])
-                    .observe(size as f64);
-            }
-        }
-
-        let (state, reason, error_code) = match output.status() {
-            TransactionStatus::Keep(execution_status) => match execution_status {
-                ExecutionStatus::Success => ("keep_success", "", "".to_string()),
-                ExecutionStatus::OutOfGas => ("keep_rejected", "OutOfGas", "error".to_string()),
-                ExecutionStatus::MoveAbort { info, .. } => (
-                    "keep_rejected",
-                    "MoveAbort",
-                    if detailed_counters {
-                        info.as_ref()
-                            .map(|v| v.reason_name.to_lowercase())
-                            .unwrap_or_else(|| "none".to_string())
-                    } else {
-                        "error".to_string()
-                    },
-                ),
-                ExecutionStatus::ExecutionFailure { .. } => {
-                    ("keep_rejected", "ExecutionFailure", "error".to_string())
-                },
-                ExecutionStatus::MiscellaneousError(e) => (
-                    "keep_rejected",
-                    "MiscellaneousError",
-                    if detailed_counters {
-                        e.map(|v| format!("{:?}", v).to_lowercase())
-                            .unwrap_or_else(|| "none".to_string())
-                    } else {
-                        "error".to_string()
-                    },
-                ),
-            },
-            TransactionStatus::Discard(discard_status_code) => {
-                (
-                    // Specialize duplicate txns for alerts
-                    if *discard_status_code == StatusCode::SEQUENCE_NUMBER_TOO_OLD {
-                        "discard_sequence_number_too_old"
-                    } else if *discard_status_code == StatusCode::SEQUENCE_NUMBER_TOO_NEW {
-                        "discard_sequence_number_too_new"
-                    } else if *discard_status_code == StatusCode::TRANSACTION_EXPIRED {
-                        "discard_transaction_expired"
-                    } else {
-                        // Only log if it is an interesting discard
-                        sample!(
-                            SampleRate::Duration(Duration::from_secs(15)),
-                            warn!(
-                                "[sampled] Txn being discarded is {:?} with status code {:?}",
-                                txn, discard_status_code
-                            )
-                        );
-                        "discard"
-                    },
-                    "error_code",
-                    if detailed_counters {
-                        format!("{:?}", discard_status_code).to_lowercase()
-                    } else {
-                        "error".to_string()
-                    },
-                )
-            },
-            TransactionStatus::Retry => ("retry", "", "".to_string()),
-        };
-
-        let kind = match txn.get_transaction() {
-            Some(Transaction::UserTransaction(_)) => "user_transaction",
-            Some(Transaction::GenesisTransaction(_)) => "genesis",
-            Some(Transaction::BlockMetadata(_)) => "block_metadata",
-            Some(Transaction::BlockMetadataExt(_)) => "block_metadata_ext",
-            Some(Transaction::StateCheckpoint(_)) => "state_checkpoint",
-            Some(Transaction::BlockEpilogue(_)) => "block_epilogue",
-            Some(Transaction::ValidatorTransaction(_)) => "validator_transaction",
-            None => "unknown",
-        };
-
-        metrics::PROCESSED_TXNS_COUNT
-            .with_label_values(&[process_type, kind, state])
-            .inc();
-
-        if !error_code.is_empty() {
-            metrics::PROCESSED_FAILED_TXNS_REASON_COUNT
-                .with_label_values(&[
-                    detailed_counters_label,
-                    process_type,
-                    state,
-                    reason,
-                    &error_code,
-                ])
-                .inc();
-        }
-
-        if let Some(Transaction::UserTransaction(user_txn)) = txn.get_transaction() {
-            if detailed_counters {
-                let mut signature_count = 0;
-                let account_authenticators = user_txn.authenticator_ref().all_signers();
-                for account_authenticator in account_authenticators {
-                    match account_authenticator {
-                        AccountAuthenticator::Ed25519 { .. } => {
-                            signature_count += 1;
-                            metrics::PROCESSED_TXNS_AUTHENTICATOR
-                                .with_label_values(&[process_type, "Ed25519"])
-                                .inc();
-                        },
-                        AccountAuthenticator::MultiEd25519 { signature, .. } => {
-                            let count = signature.signatures().len();
-                            signature_count += count;
-                            metrics::PROCESSED_TXNS_AUTHENTICATOR
-                                .with_label_values(&[process_type, "Ed25519_in_MultiEd25519"])
-                                .inc_by(count as u64);
-                        },
-                        AccountAuthenticator::SingleKey { authenticator } => {
-                            signature_count += 1;
-                            metrics::PROCESSED_TXNS_AUTHENTICATOR
-                                .with_label_values(&[
-                                    process_type,
-                                    &format!("{}_in_SingleKey", authenticator.signature().name()),
-                                ])
-                                .inc();
-                        },
-                        AccountAuthenticator::MultiKey { authenticator } => {
-                            for (_, signature) in authenticator.signatures() {
-                                signature_count += 1;
-                                metrics::PROCESSED_TXNS_AUTHENTICATOR
-                                    .with_label_values(&[
-                                        process_type,
-                                        &format!("{}_in_MultiKey", signature.name()),
-                                    ])
-                                    .inc();
-                            }
-                        },
-                        AccountAuthenticator::NoAccountAuthenticator => {
-                            metrics::PROCESSED_TXNS_AUTHENTICATOR
-                                .with_label_values(&[process_type, "NoAccountAuthenticator"])
-                                .inc();
-                        },
-                    };
-                }
-
-                metrics::PROCESSED_TXNS_NUM_AUTHENTICATORS
-                    .with_label_values(&[process_type])
-                    .observe(signature_count as f64);
-            }
-
-            match user_txn.payload() {
-                aptos_types::transaction::TransactionPayload::Script(_script) => {
-                    metrics::PROCESSED_USER_TXNS_BY_PAYLOAD
-                        .with_label_values(&[process_type, "script", state])
-                        .inc();
-                },
-                aptos_types::transaction::TransactionPayload::EntryFunction(function) => {
-                    metrics::PROCESSED_USER_TXNS_BY_PAYLOAD
-                        .with_label_values(&[process_type, "function", state])
-                        .inc();
-
-                    let is_core = function.module().address() == &CORE_CODE_ADDRESS;
-                    metrics::PROCESSED_USER_TXNS_ENTRY_FUNCTION_BY_MODULE
-                        .with_label_values(&[
-                            detailed_counters_label,
-                            process_type,
-                            if is_core { "core" } else { "user" },
-                            if detailed_counters {
-                                function.module().name().as_str()
-                            } else if is_core {
-                                "core_module"
-                            } else {
-                                "user_module"
-                            },
-                            state,
-                        ])
-                        .inc();
-                    if is_core && detailed_counters {
-                        metrics::PROCESSED_USER_TXNS_ENTRY_FUNCTION_BY_CORE_METHOD
-                            .with_label_values(&[
-                                process_type,
-                                function.module().name().as_str(),
-                                function.function().as_str(),
-                                state,
-                            ])
-                            .inc();
-                    }
-                },
-                aptos_types::transaction::TransactionPayload::Multisig(_) => {
-                    metrics::PROCESSED_USER_TXNS_BY_PAYLOAD
-                        .with_label_values(&[process_type, "multisig", state])
-                        .inc();
-                },
-
-                // Deprecated.
-                aptos_types::transaction::TransactionPayload::ModuleBundle(_) => {
-                    metrics::PROCESSED_USER_TXNS_BY_PAYLOAD
-                        .with_label_values(&[process_type, "deprecated_module_bundle", state])
-                        .inc();
-                },
-            }
-        }
-
-        for event in output.events() {
-            let (is_core, creation_number) = match event {
-                ContractEvent::V1(v1) => (
-                    v1.key().get_creator_address() == CORE_CODE_ADDRESS,
-                    if detailed_counters {
-                        v1.key().get_creation_number().to_string()
-                    } else {
-                        "event".to_string()
-                    },
-                ),
-                ContractEvent::V2(_v2) => (false, "event".to_string()),
-            };
-            metrics::PROCESSED_USER_TXNS_CORE_EVENTS
-                .with_label_values(&[
-                    detailed_counters_label,
-                    process_type,
-                    if is_core { "core" } else { "user" },
-                    &creation_number,
-                ])
-                .inc();
-        }
     }
 }
